+++ src/main/java/com/el/ally/invoicify/models/InvoiceView.java
@@ -2,7 +2,7 @@ package com.el.ally.invoicify.models;
 
 public class InvoiceView {
 	private String invoiceDescription;
-	Long recordIds[];
+	int recordIds[];
 	
 	public String getInvoiceDescription() {
 		return invoiceDescription;
@@ -10,10 +10,10 @@ public class InvoiceView {
 	public void setInvoiceDescription(String invoiceDescription) {
 		this.invoiceDescription = invoiceDescription;
 	}
-	public Long[] getRecordIds() {
+	public int[] getRecordIds() {
 		return recordIds;
 	}
-	public void setRecordIds(Long[] recordIds) {
+	public void setRecordIds(int[] recordIds) {
 		this.recordIds = recordIds;
 	}
